[shortcut-fe] add option to skip to bridge device on ingress to bridge device
Packets arriving on a bridge device trigger two calls to
netif_receive_skb(). The first time via the physical device, and the
second time as the actual bridge device

This just adds an option to skip straight to the bridge device on
ingress to skip quite a bit of overhead from the Linux networking
stack. We also update bridge stats if this is occuring.

Skipping on ingress isn't as big a deal as egress since we don't
skip qdiscs in this path

Change-Id: Ia5c1fb410919373e928c69677ed34d47831777f6
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>