[shortcut-fe] only check for a multicast ip address, not ethernet address
We started checking for multicast ip addresses in addition to mac
addresses, but we don't need to do both. Drop checking the mac addr.

This fixes an issue where ppp interfaces (which don't really have
mac addresses) are reporting invalid mac addresses which are
sometimes interpreted as multicast mac addresses causing SFE to
not offload

Change-Id: Idc159f7b1b4c30ff3a52657d034cde84dd024f75
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>