reduce the buffer size of the genl broadcast msgs
NLMSG_GOODSIZE resolves to a PAGE_SIZE when allocating the skb
associated with netlink messages.  Further, it is the skb's
truesize, not the actual used len, that is used when accounting the
socket rcvbuffer size.

This change should allow us to fit more messages into a socket's
receive queue before dropping messages due to full buffer.

Change-Id: I892dc08517e158b82d09726ae6b76cffefecb263
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>