do not auto offload connections with non-zero mark
the idea is that we only want to auto offload if userspace does not
have enough cycles to keep up with detection.  if userspace is
receiving enough cycles to mark flows, we assume that it is capable
of offload.

this fixes a specific scenario, where StreamBoost wants to continue
to monitor a connection for characteristic changes, but
fast-classifier auto offloads the connection, causing StreamBoost to
miss the change later and making it look like a misdetect.

Change-Id: Ib48e8dbbfa557b192f976f4de57a0fe5cd8f2627
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>