handle reversed tuple in offload netlink message
connections in the wan-to-lan direction were not being offloaded.

due to a quirk in the way StreamBoost classifies traffic, a connection
originating from the wan is instead represented as if it originated
from the lan.  because of this, StreamBoost sends an offload message
with the src and dst information reversed, which causes a miss when we
lookup the connection in the hash table and ultimately results in a
failure to offload.  we were saved from this in the past due to the
auto-offload feature, but with commit 020e1a81 we no longer automatically
offload a connection with a non-zero connmark.

with this patch, we try to recognize and recover from the situation by
reversing the tuple and performing a second hash table lookup.

Change-Id: I3a39f7e6bdf1ce84df7150f129ac39ac8086565a
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>