Revert "do not auto offload connections with non-zero mark"
This reverts commit 020e1a810702bfed0a75734d3bacdcf9bb79a0bb.

It turns out that this change breaks some throughput tests if
the packet source has any kind of delay between the first
packet of a connection and the ramped-up througput.  In this
case, StreamBoost receives enough cycles to set the default mark
on the connection, but then the packet flood kicks in and
userspace no longer receives any cycles and we end up in a
condition where most (if not all) connections have a default
mark but StreamBoost is no longer capable of classifying and
offloading.

iperf can exhibit the behavior, but it's only consistent if many
connections are used.  For example, "iperf -P 8" seems to work fine
and we can see good offloaded throughput numbers, but "iperf -P 16"
results in a failure to offload any connections and poor throughput.

Change-Id: I311307c69670ffd2aa8da6cd23bce358db70850c
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>